requirements.txt
gitpython
yaml
conda-execute
